# redfish/openapi/redfish-openapi.yaml
openapi: 3.0.0
info:
  title: Redfish API
  version: 1.0.0
  description: Merged Redfish API specification
paths:
  /redfish/v1/:
    get:
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceRoot_ServiceRoot'
          description: The response contains a representation of the ServiceRoot resource
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RedfishError'
          description: Error condition
      security:
      - {}
...
  /redfish/v1/SessionService:
    get:
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionService_SessionService'
          description: The response contains a representation of the SessionService
            resource
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RedfishError'
          description: Error condition
      security:
      - BasicAuth: []
    patch:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionService_SessionService'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionService_SessionService'
          description: The response contains a representation of the SessionService
            resource
        '204':
          description: Success, but no response data
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RedfishError'
          description: Error condition
      security:
      - BasicAuth: []
    put:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionService_SessionService'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionService_SessionService'
          description: The response contains a representation of the SessionService
            resource
        '204':
          description: Success, but no response data
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RedfishError'
          description: Error condition
      security:
      - BasicAuth: []
  /redfish/v1/SessionService/Sessions:
    get:
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionCollection_SessionCollection'
          description: The response contains a representation of the SessionCollection
            resource
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RedfishError'
          description: Error condition
      security:
      - BasicAuth: []
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Session_Session'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session_Session'
          description: A resource of type Session has been created
        '204':
          description: Success, but no response data
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RedfishError'
          description: Error condition
      security:
      - BasicAuth: []
...
  /redfish/v1/Systems:
    get:
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComputerSystemCollection_ComputerSystemCollection'
          description: The response contains a representation of the ComputerSystemCollection
            resource
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RedfishError'
          description: Error condition
      security:
      - BasicAuth: []
...
components:
  schemas:
...
    ComputerSystemCollection_ComputerSystemCollection:
      additionalProperties: false
      description: The collection of `ComputerSystem` resource instances.
      properties:
        '@odata.context':
          $ref: '#/components/schemas/odata-v4_context'
        '@odata.etag':
          $ref: '#/components/schemas/odata-v4_etag'
        '@odata.id':
          $ref: '#/components/schemas/odata-v4_id'
        '@odata.type':
          $ref: '#/components/schemas/odata-v4_type'
        Description:
          oneOf:
          - $ref: '#/components/schemas/Resource_Description'
          - enum:
            - null
          readOnly: true
        Members:
          description: The members of this collection.
          items:
            $ref: '#/components/schemas/odata-v4_idRef'
          readOnly: true
          type: array
          x-longDescription: This property shall contain an array of links to the
            members of this collection.
        Members@odata.count:
          $ref: '#/components/schemas/odata-v4_count'
        Members@odata.nextLink:
          $ref: '#/components/schemas/odata-v4_nextLink'
        Name:
          $ref: '#/components/schemas/Resource_Name'
          readOnly: true
        Oem:
          $ref: '#/components/schemas/Resource_Oem'
          description: The OEM extension property.
          x-longDescription: This property shall contain the OEM extensions.  All
            values for properties contained in this object shall conform to the Redfish
            Specification-described requirements.
      required:
      - Members
      - Members@odata.count
      - '@odata.id'
      - '@odata.type'
      - Name
      type: object
      x-longDescription: This resource shall represent a resource collection of `ComputerSystem`
        instances for a Redfish implementation.
      x-patternProperties:
        ^([a-zA-Z_][a-zA-Z0-9_]*)?@(odata|Redfish|Message)\.[a-zA-Z_][a-zA-Z0-9_]*$:
          description: This property shall specify a valid odata or Redfish property.
...
    RedfishError:
      description: The error payload from a Redfish service.
      properties:
        error:
          description: The properties that describe an error from a Redfish service.
          properties:
            '@Message.ExtendedInfo':
              description: An array of messages describing one or more error messages.
              items:
                $ref: '#/components/schemas/Message_Message'
              type: array
              x-longDescription: This property shall be an array of message objects
                describing one or more error messages.
            code:
              description: A string indicating a specific `MessageId` from a message
                registry.
              readOnly: true
              type: string
              x-longDescription: This property shall contain a string indicating a
                specific `MessageId` from a message registry.
            message:
              description: A human-readable error message corresponding to the message
                in a message registry.
              readOnly: true
              type: string
              x-longDescription: This property shall contain a human-readable error
                message corresponding to the message in a message registry.
          required:
          - code
          - message
          type: object
          x-longDescription: The Redfish Specification-described type shall contain
            properties that describe an error from a Redfish service.
      required:
      - error
      type: object
      x-longDescription: The Redfish Specification-described type shall contain an
        error payload from a Redfish service.
...
    ServiceRoot_ServiceRoot:
      additionalProperties: false
      description: The `ServiceRoot` schema describes the root of the Redfish service,
        located at the '/redfish/v1' URI.  All other resources accessible through
        the Redfish interface on this device are linked directly or indirectly from
        the service root.
      properties:
        '@odata.context':
          $ref: '#/components/schemas/odata-v4_context'
        '@odata.id':
          $ref: '#/components/schemas/odata-v4_id'
        '@odata.type':
          $ref: '#/components/schemas/odata-v4_type'
        Description:
          oneOf:
          - $ref: '#/components/schemas/Resource_Description'
          - enum:
            - null
          readOnly: true
        Id:
          $ref: '#/components/schemas/Resource_Id'
          readOnly: true
        Links:
          $ref: '#/components/schemas/ServiceRoot_Links'
          description: The links to other resources that are related to this resource.
          nullable: true
          x-longDescription: This property shall contain links to resources that are
            related to but are not contained by, or subordinate to, this resource.
        Name:
          $ref: '#/components/schemas/Resource_Name'
          readOnly: true
        Product:
          description: The product associated with this Redfish service.
          nullable: true
          readOnly: true
          type: string
          x-longDescription: This property shall include the name of the product represented
            by this Redfish service.
          x-versionAdded: v1_3_0
        RedfishVersion:
          description: The version of the Redfish service.
          pattern: ^\d+\.\d+\.\d+$
          readOnly: true
          type: string
          x-longDescription: This property shall represent the Redfish protocol version,
            as specified in the 'Protocol version' clause of the Redfish Specification,
            to which this service conforms.
        Systems:
          $ref: '#/components/schemas/odata-v4_idRef'
          description: The link to a collection of systems.
          readOnly: true
          x-longDescription: This property shall contain a link to a resource collection
            of type `ComputerSystemCollection`.
        UUID:
          description: Unique identifier for a service instance.  When SSDP is used,
            this value contains the same UUID returned in an HTTP `200 OK` response
            from an SSDP `M-SEARCH` request during discovery.
          nullable: true
          pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
          readOnly: true
          type: string
          x-longDescription: This property shall contain the identifier of the Redfish
            service instance.  If SSDP is used, this value shall contain the same
            UUID returned in an HTTP `200 OK` response from an SSDP `M-SEARCH` request
            during discovery.  RFC4122 describes methods to use to create a UUID value.  The
            value should be considered to be opaque.  Client software should only
            treat the overall value as a universally unique identifier and should
            not interpret any subfields within the UUID.
        Vendor:
          description: The vendor or manufacturer associated with this Redfish service.
          nullable: true
          readOnly: true
          type: string
          x-longDescription: This property shall include the name of the manufacturer
            or vendor represented by this Redfish service.  If this property is supported,
            the vendor name shall not be included in the `Product` property value.
          x-versionAdded: v1_5_0
      required:
      - '@odata.id'
      - '@odata.type'
      - Id
      - Name
      type: object
      x-longDescription: This resource shall represent the root of the Redfish service.
      x-patternProperties:
        ^([a-zA-Z_][a-zA-Z0-9_]*)?@(odata|Redfish|Message)\.[a-zA-Z_][a-zA-Z0-9_]*$:
          description: This property shall specify a valid odata or Redfish property.
    SessionCollection_SessionCollection:
      additionalProperties: false
      description: The collection of `Session` resource instances.
      properties:
        '@odata.context':
          $ref: '#/components/schemas/odata-v4_context'
        '@odata.etag':
          $ref: '#/components/schemas/odata-v4_etag'
        '@odata.id':
          $ref: '#/components/schemas/odata-v4_id'
        '@odata.type':
          $ref: '#/components/schemas/odata-v4_type'
        Description:
          oneOf:
          - $ref: '#/components/schemas/Resource_Description'
          - enum:
            - null
          readOnly: true
        Members:
          description: The members of this collection.
          items:
            $ref: '#/components/schemas/odata-v4_idRef'
          readOnly: true
          type: array
          x-longDescription: This property shall contain an array of links to the
            members of this collection.
        Members@odata.count:
          $ref: '#/components/schemas/odata-v4_count'
        Members@odata.nextLink:
          $ref: '#/components/schemas/odata-v4_nextLink'
        Name:
          $ref: '#/components/schemas/Resource_Name'
          readOnly: true
        Oem:
          $ref: '#/components/schemas/Resource_Oem'
          description: The OEM extension property.
          x-longDescription: This property shall contain the OEM extensions.  All
            values for properties contained in this object shall conform to the Redfish
            Specification-described requirements.
      required:
      - Members
      - Members@odata.count
      - '@odata.id'
      - '@odata.type'
      - Name
      type: object
      x-longDescription: This resource shall represent a resource collection of `Session`
        instances for a Redfish implementation.
      x-patternProperties:
        ^([a-zA-Z_][a-zA-Z0-9_]*)?@(odata|Redfish|Message)\.[a-zA-Z_][a-zA-Z0-9_]*$:
          description: This property shall specify a valid odata or Redfish property.
...
    SessionService_SessionService:
      additionalProperties: false
      description: The `SessionService` schema describes the session service and its
        properties, with links to the actual list of sessions.
      properties:
        '@odata.context':
          $ref: '#/components/schemas/odata-v4_context'
        '@odata.etag':
          $ref: '#/components/schemas/odata-v4_etag'
        '@odata.id':
          $ref: '#/components/schemas/odata-v4_id'
        '@odata.type':
          $ref: '#/components/schemas/odata-v4_type'
        AbsoluteSessionTimeout:
          description: The maximum number of seconds that a session is open before
            the service closes the session regardless of activity.
          format: int64
          maximum: 86400
          minimum: 30
          readOnly: false
          type: integer
          x-longDescription: This property shall contain the maximum number of seconds
            that a session is open before the service closes the session regardless
            of activity.
          x-units: s
          x-versionAdded: v1_2_0
        AbsoluteSessionTimeoutEnabled:
          description: An indication of whether an absolute session timeout is applied
            to sessions.
          readOnly: false
          type: boolean
          x-longDescription: This property shall indicate whether an absolute session
            timeout is applied to sessions.  If `true`, the service shall close sessions
            that are open for the number of seconds specified by the `AbsoluteSessionTimeout`
            property regardless of session activity.  If `false` or if this property
            is not present, the service shall not apply an absolute session timeout.
          x-versionAdded: v1_2_0
        Actions:
          $ref: '#/components/schemas/SessionService_Actions'
          description: The available actions for this resource.
          x-longDescription: This property shall contain the available actions for
            this resource.
          x-versionAdded: v1_1_0
        Description:
          oneOf:
          - $ref: '#/components/schemas/Resource_Description'
          - enum:
            - null
          readOnly: true
        Id:
          $ref: '#/components/schemas/Resource_Id'
          readOnly: true
        Name:
          $ref: '#/components/schemas/Resource_Name'
          readOnly: true
        ServiceEnabled:
          description: An indication of whether this service is enabled.  If `true`,
            this service is enabled.  If `false`, it is disabled, and new sessions
            cannot be created, old sessions cannot be deleted, and established sessions
            can continue operating.
          nullable: true
          readOnly: false
          type: boolean
          x-longDescription: This property shall indicate whether this service is
            enabled.  If `true`, this service is enabled.  If `false`, it is disabled,
            and new sessions shall not be created, old sessions shall not be deleted,
            and established sessions can continue operating.
        SessionTimeout:
          description: The number of seconds of inactivity that a session can have
            before the session service closes the session due to inactivity.
          format: int64
          maximum: 86400
          minimum: 30
          readOnly: false
          type: integer
          x-longDescription: This property shall contain the threshold of time in
            seconds between requests on a specific session at which point the session
            service shall close the session due to inactivity.  The session service
            shall support any value between the schema-specified minimum and maximum
            terms.
          x-units: s
        Sessions:
          $ref: '#/components/schemas/odata-v4_idRef'
          description: The link to a collection of sessions.
          readOnly: true
          x-longDescription: This property shall contain a link to a resource collection
            of type `SessionCollection`.
        Status:
          $ref: '#/components/schemas/Resource_Status'
          description: The status and health of the resource and its subordinate or
            dependent resources.
          x-longDescription: This property shall contain any status or health properties
            of the resource.
      required:
      - '@odata.id'
      - '@odata.type'
      - Id
      - Name
      type: object
      x-longDescription: This resource contains the session service properties for
        a Redfish implementation.
      x-patternProperties:
        ^([a-zA-Z_][a-zA-Z0-9_]*)?@(odata|Redfish|Message)\.[a-zA-Z_][a-zA-Z0-9_]*$:
          description: This property shall specify a valid odata or Redfish property.
...
    Session_Session:
      additionalProperties: false
      description: The `Session` resource describes a single connection (session)
        between a client and a Redfish service instance.
      properties:
        '@odata.context':
          $ref: '#/components/schemas/odata-v4_context'
        '@odata.etag':
          $ref: '#/components/schemas/odata-v4_etag'
        '@odata.id':
          $ref: '#/components/schemas/odata-v4_id'
        '@odata.type':
          $ref: '#/components/schemas/odata-v4_type'
        Actions:
          $ref: '#/components/schemas/Session_Actions'
          description: The available actions for this resource.
          x-longDescription: This property shall contain the available actions for
            this resource.
          x-versionAdded: v1_1_0
        ClientOriginIPAddress:
          description: The IP address of the client that created the session.
          nullable: true
          readOnly: true
          type: string
          x-longDescription: This property shall contain the IP address of the client
            that created the session.
          x-versionAdded: v1_3_0
        Context:
          description: A client-supplied string that is stored with the session.
          nullable: true
          readOnly: true
          type: string
          x-longDescription: This property shall contain a client-supplied context
            that remains with the session through the session's lifetime.
          x-versionAdded: v1_5_0
        CreatedTime:
          description: The date and time when the session was created.
          format: date-time
          nullable: true
          readOnly: true
          type: string
          x-longDescription: This property shall contain the date and time when the
            session was created.
          x-versionAdded: v1_4_0
        Description:
          oneOf:
          - $ref: '#/components/schemas/Resource_Description'
          - enum:
            - null
          readOnly: true
        ExpirationTime:
          description: The date and time when the session expires regardless of session
            activity.
          format: date-time
          readOnly: true
          type: string
          x-longDescription: This property shall contain the date and time when the
            session expires regardless of session activity.  The service shall delete
            this resource when the expiration time is reached.  If this property is
            not present, the session does not expire based on an absolute time.
          x-versionAdded: v1_8_0
        Id:
          $ref: '#/components/schemas/Resource_Id'
          readOnly: true
        Links:
          $ref: '#/components/schemas/Session_Links'
          description: The links to other resources that are related to this resource.
          x-longDescription: This property shall contain links to resources that are
            related to but are not contained by, or subordinate to, this resource.
          x-versionAdded: v1_7_0
        Name:
          $ref: '#/components/schemas/Resource_Name'
          readOnly: true
        Password:
          description: The password for this session.  The value is `null` in responses.
          nullable: true
          readOnly: true
          type: string
          writeOnly: false
          x-longDescription: This property shall contain the password for this session.  The
            value shall be `null` in responses.  When creating a session through a
            Redfish host interface using an `AuthNone` role, the property shall contain
            an empty string in the request body.
        Roles:
          description: The Redfish roles that contain the privileges of this session.
          items:
            nullable: true
            type: string
          readOnly: true
          type: array
          x-longDescription: This property shall contain the Redfish roles that contain
            the privileges of this session.
          x-versionAdded: v1_7_0
        SessionType:
          description: The active session type.
          oneOf:
          - $ref: '#/components/schemas/Session_SessionTypes'
          - enum:
            - null
          readOnly: true
          x-longDescription: This property shall represent the type of session that
            is currently active.
          x-versionAdded: v1_2_0
        Token:
          description: The multi-factor authentication token for this session.  The
            value is `null` in responses.
          nullable: true
          readOnly: true
          type: string
          writeOnly: false
          x-longDescription: This property shall contain the multi-factor authentication
            token for this session.  The value shall be `null` in responses.
          x-versionAdded: v1_6_0
        UserName:
          description: The username for the account for this session.
          nullable: true
          readOnly: true
          type: string
          x-longDescription: This property shall contain the username that matches
            an account recognized by the account service.  When a creating a session
            through a Redfish host interface using an `AuthNone` role, the property
            shall contain an empty string in the request body.
      required:
      - '@odata.id'
      - '@odata.type'
      - Id
      - Name
      type: object
      x-longDescription: This resource shall represent a session for a Redfish implementation.
      x-patternProperties:
        ^([a-zA-Z_][a-zA-Z0-9_]*)?@(odata|Redfish|Message)\.[a-zA-Z_][a-zA-Z0-9_]*$:
          description: This property shall specify a valid odata or Redfish property.
      x-requiredOnCreate:
      - UserName
      - Password
...
    odata-v4_context:
      description: The OData description of a payload.
      format: uri-reference
      readOnly: true
      type: string
      x-longDescription: The value of this property shall be the context URL that
        describes the resource according to OData-Protocol and shall be of the form
        defined in the Redfish specification.
...
    odata-v4_id:
      description: The unique identifier for a resource.
      format: uri-reference
      readOnly: true
      type: string
      x-longDescription: The value of this property shall be the unique identifier
        for the resource and it shall be of the form defined in the Redfish specification.
...
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
      description: HTTP Basic Authentication for Redfish API
security:
- BasicAuth: []
- {}
...